% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{BCharge}
\alias{BCharge}
\title{BCharge
Finds charge on any buffer given activity coefficients (FF), H, KA,tot}
\usage{
BCharge(TOT, FF, KA, H)
}
\arguments{
\item{TOT}{vectore of buffer concentrations}

\item{FF}{vactor of activity coefficients}

\item{KA}{vector of dissociation constants}

\item{H}{proton concentration}
}
\description{
BCharge
Finds charge on any buffer given activity coefficients (FF), H, KA,tot
}
