% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{Bspecif}
\alias{Bspecif}
\title{Bspecif
Finds the partition of buffer species}
\usage{
Bspecif(TTOT, pH, KA, FF)
}
\arguments{
\item{TTOT, pH}{Total buffer concentration and pH}

\item{KA, FF}{vectors of dissociations and activivty coefficients}
}
\description{
Bspecif
Finds the partition of buffer species
}
