% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{CMB}
\alias{CMB}
\title{CMB
Generic function to find charge}
\usage{
CMB(Na, K, Cl, Ca, Mg, Lact, H, WA, FF, TOT, PCO2, Alb)
}
\arguments{
\item{Na, K, Cl, Ca, Mg, Lact}{Strong ions for the charge balance}

\item{WA, TOT}{Buffer specifications and total concentrations}

\item{FF, H, PCO2, Alb}{Activity coefficients, H, PCO2 and albumin}
}
\description{
CMB
Generic function to find charge
}
