% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{CO2specif}
\alias{CO2specif}
\title{CO2specif - finds the partition between CO2 species}
\usage{
CO2specif(PCO2, pH, FF)
}
\arguments{
\item{PCO2, pH, FF}{CO2 tension, pH and activity coefficients}
}
\description{
CO2specif - finds the partition between CO2 species
}
