% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{GET_CH}
\alias{GET_CH}
\title{GET_CH
Calls BCharge to gather charges from a buffer}
\usage{
GET_CH(TOT, FF, WA, H)
}
\arguments{
\item{TOT}{vectore of buffer concentrations}

\item{FF}{vactor of activity coefficients}

\item{WA}{list of buffer specifications}

\item{H}{proton concentration}
}
\description{
GET_CH
Calls BCharge to gather charges from a buffer
}
