% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{getfs}
\alias{getfs}
\title{Getfs

Gets acitivity coefficiens from Davies equation.}
\usage{
getfs(z, I, A, b)
}
\arguments{
\item{z, I}{- vector of charges and ionic strength}

\item{A, b}{are parameters in Davies equation}
}
\description{
Getfs

Gets acitivity coefficiens from Davies equation.
}
