% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_acidb.R
\name{getI}
\alias{getI}
\title{GetI - finds the ionic strength}
\usage{
getI(CONC, VALS)
}
\arguments{
\item{CONC, VALS}{concentrations and valencies}
}
\description{
GetI - finds the ionic strength
}
