% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_acidb.R
\name{pH_general}
\alias{pH_general}
\title{Find equilibrium parameters}
\usage{
pH_general(Na = 0, K = 0, Cl = 0, Lact = 0, Ca = 0, Mg = 0,
  PCO2 = 0, Alb = 0, WA = list(), TOT, A = A, b = b)
}
\arguments{
\item{Na, K, Cl, Lact, Ca, Mg}{Strong ions inputs, "0" if not present.}

\item{PCO2, Alb}{If present PCO2 and Albumin are handled - "0" if not present.}

\item{WA}{List of lists with names of all buffers in first list, and all pK
values from smallest to highest for each buffer in second list.}

\item{TOT}{Vector of total buffer concentrations.}

\item{A, b}{Constants from Davies equation.}
}
\value{
Returns pH: pH as -log10 of direct H from charge balance
  minimization after correction for ionic strength, KpH: -log10 of that H
  multiplied by F1 (activity coeff for charge=1), UNCOR_pH: pH with all
  activity coeffs = 1,II: final ionic strength, FII: ionic strength before
  optimizing activity coefficients, i: number of iterations in optimization,
  SPECS: vector of all buffer species, FSPECS: species distribution before
  optimizing activity coefficients.
}
\description{
Find equilibrium parameters
}
